In this PR, I'm making rule actions run ~500ms faster (20ms to 1s faster) by avoiding to wait for a refresh to happen when deleting the `action_task_params` saved-object. This refresh made the delete operation run anywhere from ~20ms to 1s accounting for a large part of the action execution time.